@@ -22,6 +22,6 @@
     "swagger-ui-express": "^4.1.6"
   },
   "engines": {
-    "node": "15.x"
+    "node": "18.x"
   }
 }
